// Qanat.Web/src/app/shared/generated/model/meter-link-display-dto.ts
/**
 * Qanat.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class MeterLinkDisplayDto { 
    MeterID?: number;
    LinkDisplay?: string;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface MeterLinkDisplayDtoForm { 
    MeterID?: FormControl<number>;
    LinkDisplay?: FormControl<string>;
}

export class MeterLinkDisplayDtoFormControls { 
    public static MeterID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static LinkDisplay = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}